[Silabs] Move Wi-Fi interface to src/platform from examples/platform (#36477)
* Restructure platform files

* Move all files to src/platform

* Build the wf200 in the new wifi structure

* rename directory

* Build the rs9116 in the new wifi structure

* Build the SiWx917 NCP in the new wifi structure

* Build the SiWx917 SoC in the new wifi structure

* Delete wfx_host_events header in favor of the new abstraction structure

* remove unnecessary includes in gn files and the cflag changes

* restyle

* add todo comment

* remove cflag inclusion

* add missing incllude

* Add missing includes

* Restyled by clang-format

* Move wifi_args.gni file to the wifi directory

* Rename Wf200WifiInterface to WifiInterface

* Rename SiWxWifiInterface to WifiInterface

* Rename Rsi91xWifiInterface.cpp to WifiInterface.cpp

---------
