+++ examples/build_overrides/efr32_sdk.gni
@@ -14,5 +14,5 @@
 
 declare_args() {
   # Root directory for erf32 SDK.
-  efr32_sdk_build_root = "//third_party/connectedhomeip/third_party/silabs"
+  silabs_sdk_build_root = "//third_party/connectedhomeip/third_party/silabs"
 }
