+++ examples/platform/silabs/efr32/rs911x/rs911x.gni
@@ -1,6 +1,6 @@
 import("//build_overrides/chip.gni")
 import("//build_overrides/efr32_sdk.gni")
-import("${efr32_sdk_build_root}/efr32_sdk.gni")
+import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 rs911x_src_plat = [
   "${examples_plat_dir}/rs911x/rsi_if.cpp",
