+++ examples/platform/silabs/efr32/rs911x/rsi_if.cpp
@@ -50,6 +50,7 @@ extern "C" {
 #endif
 
 #include "dhcp_client.h"
+#include "ethernetif.h"
 #include "lwip/nd6.h"
 #include "silabs_utils.h"
 #include "wfx_rsi.h"
@@ -109,7 +110,7 @@ static void DHCPTimerEventHandler(void * arg)
 {
     WfxEvent_t event;
     event.eventType = WFX_EVT_DHCP_POLL;
-    WfxPostEvent(&event);
+    sl_matter_wifi_post_event(&event);
 }
 
 static void CancelDHCPTimer(void)
@@ -226,14 +227,14 @@ int32_t wfx_rsi_reset_count(void)
 }
 
 /******************************************************************
- * @fn   wfx_rsi_disconnect(void)
+ * @fn   sl_wifi_platform_disconnect(void)
  * @brief
  *       Getting the driver disconnect status
  * @param[in] None
  * @return
  *        status
  *********************************************************************/
-int32_t wfx_rsi_disconnect(void)
+int32_t sl_wifi_platform_disconnect(void)
 {
     return rsi_wlan_disconnect();
 }
@@ -301,7 +302,7 @@ static void wfx_rsi_join_cb(uint16_t status, const uint8_t * buf, const uint16_t
     ChipLogProgress(DeviceLayer, "wfx_rsi_join_cb: success");
     memset(&temp_reset, 0, sizeof(wfx_wifi_scan_ext_t));
     WfxEvent.eventType = WFX_EVT_STA_CONN;
-    WfxPostEvent(&WfxEvent);
+    sl_matter_wifi_post_event(&WfxEvent);
     wfx_rsi.join_retries = 0;
 }
 
@@ -322,7 +323,7 @@ static void wfx_rsi_join_fail_cb(uint16_t status, uint8_t * buf, uint32_t len)
     wfx_rsi.join_retries += 1;
     wfx_rsi.dev_state &= ~(WFX_RSI_ST_STA_CONNECTING | WFX_RSI_ST_STA_CONNECTED);
     WfxEvent.eventType = WFX_EVT_STA_START_JOIN;
-    WfxPostEvent(&WfxEvent);
+    sl_matter_wifi_post_event(&WfxEvent);
 }
 /*************************************************************************************
  * @fn  wfx_rsi_wlan_pkt_cb(uint16_t status, uint8_t *buf, uint32_t len)
@@ -344,31 +345,30 @@ static void wfx_rsi_wlan_pkt_cb(uint16_t status, uint8_t * buf, uint32_t len)
 }
 
 /*************************************************************************************
- * @fn  static int32_t wfx_rsi_init(void)
+ * @fn  static int32_t sl_matter_wifi_init(void)
  * @brief
  *      driver initialization
  * @param[in]  None
  * @return
  *        None
  *****************************************************************************************/
-static int32_t wfx_rsi_init(void)
+static int32_t sl_matter_wifi_init(void)
 {
     int32_t status;
     uint8_t buf[RSI_RESPONSE_HOLD_BUFF_SIZE];
     extern void rsi_hal_board_init(void);
 
-    ChipLogProgress(DeviceLayer, "wfx_rsi_init: starting(HEAP_SZ = %d)", SL_HEAP_SIZE);
     //! Driver initialization
     status = rsi_driver_init(wfx_rsi_drv_buf, WFX_RSI_BUF_SZ);
     if ((status < RSI_DRIVER_STATUS) || (status > WFX_RSI_BUF_SZ))
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_driver_init failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_driver_init failed: %ld", status);
         return status;
     }
     /* ! Redpine module intialisation */
     if ((status = rsi_device_init(LOAD_NWP_FW)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_device_init failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_device_init failed: %ld", status);
         return status;
     }
     /*
@@ -377,7 +377,7 @@ static int32_t wfx_rsi_init(void)
     sDrvThread = osThreadNew(rsi_wireless_driver_task_wrapper, NULL, &kDrvTaskAttr);
     if (NULL == sDrvThread)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: failed to create task");
+        ChipLogError(DeviceLayer, "failed to create task");
         return RSI_ERROR_INVALID_PARAM;
     }
 
@@ -388,7 +388,7 @@ static int32_t wfx_rsi_init(void)
     if ((status = rsi_wireless_init(OPER_MODE_0, COEX_MODE_0)) != RSI_SUCCESS)
     {
 #endif
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_wireless_init failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_wireless_init failed: %ld", status);
         return status;
     }
 
@@ -397,7 +397,7 @@ static int32_t wfx_rsi_init(void)
      */
     if (rsi_wlan_get(RSI_FW_VERSION, buf, sizeof(buf)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_wlan_get(RSI_FW_VERSION) failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_wlan_get(RSI_FW_VERSION) failed: %ld", status);
         return status;
     }
 
@@ -415,13 +415,12 @@ static int32_t wfx_rsi_init(void)
     (void) rsi_wlan_radio_init(); /* Required so we can get MAC address */
     if ((status = rsi_wlan_get(RSI_MAC_ADDRESS, &wfx_rsi.sta_mac.octet[0], RESP_BUFF_SIZE)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_wlan_get(RSI_MAC_ADDRESS) failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_wlan_get(RSI_MAC_ADDRESS) failed: %ld", status);
         return status;
     }
 
-    ChipLogDetail(DeviceLayer, "wfx_rsi_init: MAC: %02x:%02x:%02x %02x:%02x:%02x", wfx_rsi.sta_mac.octet[0],
-                  wfx_rsi.sta_mac.octet[1], wfx_rsi.sta_mac.octet[2], wfx_rsi.sta_mac.octet[3], wfx_rsi.sta_mac.octet[4],
-                  wfx_rsi.sta_mac.octet[5]);
+    ChipLogDetail(DeviceLayer, "MAC: %02x:%02x:%02x %02x:%02x:%02x", wfx_rsi.sta_mac.octet[0], wfx_rsi.sta_mac.octet[1],
+                  wfx_rsi.sta_mac.octet[2], wfx_rsi.sta_mac.octet[3], wfx_rsi.sta_mac.octet[4], wfx_rsi.sta_mac.octet[5]);
 
     // Create the message queue
     sWifiEventQueue = osMessageQueueNew(WFX_QUEUE_SIZE, sizeof(WfxEvent_t), NULL);
@@ -443,12 +442,12 @@ static int32_t wfx_rsi_init(void)
      */
     if ((status = rsi_wlan_register_callbacks(RSI_JOIN_FAIL_CB, wfx_rsi_join_fail_cb)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_wlan_register_callbacks failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_wlan_register_callbacks failed: %ld", status);
         return status;
     }
     if ((status = rsi_wlan_register_callbacks(RSI_WLAN_DATA_RECEIVE_NOTIFY_CB, wfx_rsi_wlan_pkt_cb)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_init: rsi_wlan_register_callbacks failed: %ld", status);
+        ChipLogError(DeviceLayer, "rsi_wlan_register_callbacks failed: %ld", status);
         return status;
     }
 
@@ -457,7 +456,6 @@ static int32_t wfx_rsi_init(void)
 #endif
 
     wfx_rsi.dev_state |= WFX_RSI_ST_DEV_READY;
-    ChipLogProgress(DeviceLayer, "wfx_rsi_init: success");
     return RSI_SUCCESS;
 }
 
@@ -530,20 +528,20 @@ static void wfx_rsi_save_ap_info(void) // translation
 }
 
 /********************************************************************************************
- * @fn   static void wfx_rsi_do_join(void)
+ * @fn   static void sl_wifi_platform_join_network(void)
  * @brief
  *        Start an async Join command
  * @return
  *        None
  **********************************************************************************************/
-static void wfx_rsi_do_join(void)
+static void sl_wifi_platform_join_network(void)
 {
-    int32_t status;
+    int32_t status = SL_STATUS_OK;
     rsi_security_mode_t connect_security_mode;
 
     if (wfx_rsi.dev_state & (WFX_RSI_ST_STA_CONNECTING | WFX_RSI_ST_STA_CONNECTED))
     {
-        ChipLogProgress(DeviceLayer, "wfx_rsi_do_join: already in progress");
+        ChipLogProgress(DeviceLayer, "sl_wifi_platform_join_network: already in progress");
         return;
     }
 
@@ -566,11 +564,12 @@ static void wfx_rsi_do_join(void)
         connect_security_mode = RSI_OPEN;
         break;
     default:
-        ChipLogError(DeviceLayer, "wfx_rsi_do_join: error: unknown security type.");
+        ChipLogError(DeviceLayer, "sl_wifi_platform_join_network: error: unknown security type.");
         return;
     }
 
-    ChipLogProgress(DeviceLayer, "wfx_rsi_do_join: connecting to %s, sec=%d", &wfx_rsi.sec.ssid[0], wfx_rsi.sec.security);
+    ChipLogProgress(DeviceLayer, "sl_wifi_platform_join_network: connecting to %s, sec=%d", &wfx_rsi.sec.ssid[0],
+                    wfx_rsi.sec.security);
 
     /*
      * Join the network
@@ -582,7 +581,7 @@ static void wfx_rsi_do_join(void)
 
     if ((status = rsi_wlan_register_callbacks(RSI_JOIN_FAIL_CB, wfx_rsi_join_fail_cb)) != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_do_join: rsi_wlan_register_callbacks failed: %ld", status);
+        ChipLogError(DeviceLayer, "sl_wifi_platform_join_network: rsi_wlan_register_callbacks failed: %ld", status);
     }
 
     /* Try to connect Wifi with given Credentials
@@ -592,7 +591,8 @@ static void wfx_rsi_do_join(void)
                                          wfx_rsi_join_cb)) != RSI_SUCCESS)
     {
         wfx_rsi.dev_state &= ~WFX_RSI_ST_STA_CONNECTING;
-        ChipLogProgress(DeviceLayer, "wfx_rsi_do_join: rsi_wlan_connect_async failed: %ld on try %d", status, wfx_rsi.join_retries);
+        ChipLogProgress(DeviceLayer, "sl_wifi_platform_join_network: rsi_wlan_connect_async failed: %ld on try %d", status,
+                        wfx_rsi.join_retries);
         wfx_retry_connection(++wfx_rsi.join_retries);
     }
 }
@@ -645,7 +645,7 @@ void HandleDHCPPolling(void)
         wfx_ipv6_notify(GET_IPV6_SUCCESS);
         hasNotifiedIPV6 = true;
         event.eventType = WFX_EVT_STA_DHCP_DONE;
-        WfxPostEvent(&event);
+        sl_matter_wifi_post_event(&event);
         NotifyConnectivity();
     }
 }
@@ -665,19 +665,16 @@ void ResetDHCPNotificationFlags(void)
     hasNotifiedWifiConnectivity = false;
 
     outEvent.eventType = WFX_EVT_STA_DO_DHCP;
-    WfxPostEvent(&outEvent);
+    sl_matter_wifi_post_event(&outEvent);
 }
 
-/**
- * @brief Post the WfxEvent to tue WiFiEventQueue to be process by the wfx_rsi_task
- */
-void WfxPostEvent(WfxEvent_t * event)
+void sl_matter_wifi_post_event(WfxEvent_t * event)
 {
     sl_status_t status = osMessageQueuePut(sWifiEventQueue, event, 0, 0);
 
     if (status != osOK)
     {
-        ChipLogError(DeviceLayer, "WfxPostEvent: failed to post event with status: %ld", status);
+        ChipLogError(DeviceLayer, "sl_matter_wifi_post_event: failed to post event with status: %ld", status);
         // TODO: Handle error, requeue event depending on queue size or notify relevant task,
         // Chipdie, etc.
     }
@@ -799,7 +796,7 @@ void ProcessEvent(WfxEvent_t inEvent)
         // saving the AP related info
         wfx_rsi_save_ap_info();
         // Joining to the network
-        wfx_rsi_do_join();
+        sl_wifi_platform_join_network();
     }
     break;
     case WFX_EVT_STA_DO_DHCP: {
@@ -820,7 +817,7 @@ void ProcessEvent(WfxEvent_t inEvent)
 }
 
 /*********************************************************************************
- * @fn  void wfx_rsi_task(void *arg)
+ * @fn  void sl_matter_wifi_task(void *arg)
  * @brief
  * The main WLAN task - started by wfx_wifi_start () that interfaces with RSI.
  * The rest of RSI stuff come in call-backs.
@@ -830,20 +827,20 @@ void ProcessEvent(WfxEvent_t inEvent)
  *       None
  **********************************************************************************/
 /* ARGSUSED */
-void wfx_rsi_task(void * arg)
+void sl_matter_wifi_task(void * arg)
 {
     (void) arg;
-    uint32_t rsi_status = wfx_rsi_init();
+    uint32_t rsi_status = sl_matter_wifi_init();
     if (rsi_status != RSI_SUCCESS)
     {
-        ChipLogError(DeviceLayer, "wfx_rsi_task: wfx_rsi_init failed: %ld", rsi_status);
+        ChipLogError(DeviceLayer, "sl_matter_wifi_task: sl_matter_wifi_init failed: %ld", rsi_status);
         return;
     }
     WfxEvent_t wfxEvent;
-    wfx_lwip_start();
-    wfx_started_notify();
+    sl_matter_lwip_start();
+    sl_matter_wifi_task_started();
 
-    ChipLogProgress(DeviceLayer, "wfx_rsi_task: starting event loop");
+    ChipLogProgress(DeviceLayer, "sl_matter_wifi_task: starting event loop");
     for (;;)
     {
         osStatus_t status = osMessageQueueGet(sWifiEventQueue, &wfxEvent, NULL, osWaitForever);
@@ -853,7 +850,7 @@ void wfx_rsi_task(void * arg)
         }
         else
         {
-            ChipLogProgress(DeviceLayer, "wfx_rsi_task: get event failed: %x", status);
+            ChipLogProgress(DeviceLayer, "sl_matter_wifi_task: get event failed: %x", status);
         }
     }
 }
