--- examples/platform/silabs/efr32/rs911x/sl_wifi_if.cpp
@@ -1,1 +1,0 @@
-../../SiWx917/SiWx917/sl_wifi_if.cpp
