+++ examples/platform/silabs/efr32/wf200/wf200.gni
@@ -1,6 +1,6 @@
 import("//build_overrides/chip.gni")
 import("//build_overrides/efr32_sdk.gni")
-import("${efr32_sdk_build_root}/efr32_sdk.gni")
+import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 wf200_plat_incs = [ "${examples_plat_dir}/wf200" ]
 wf200_plat_src = [
