+++ examples/platform/silabs/provision/BUILD.gn
@@ -14,12 +14,12 @@
 
 import("//build_overrides/chip.gni")
 import("//build_overrides/efr32_sdk.gni")
-import("${efr32_sdk_build_root}/silabs_board.gni")
+import("${silabs_sdk_build_root}/silabs_board.gni")
 
 if (wifi_soc) {
-  import("${efr32_sdk_build_root}/SiWx917_sdk.gni")
+  import("${silabs_sdk_build_root}/SiWx917_sdk.gni")
 } else {
-  import("${efr32_sdk_build_root}/efr32_sdk.gni")
+  import("${silabs_sdk_build_root}/efr32_sdk.gni")
 }
 
 source_set("storage") {
