+++ src/lwip/BUILD.gn
@@ -220,7 +220,7 @@ if (current_os == "zephyr" || current_os == "mbed") {
       sources +=
           [ "${chip_root}/third_party/lwip/repo/lwip/src/apps/mdns/mdns.c" ]
     } else if (lwip_platform == "silabs") {
-      public_deps += [ "${efr32_sdk_build_root}:efr32_sdk" ]
+      public_deps += [ "${silabs_sdk_build_root}:silabs_sdk" ]
 
       sources += [
         "${lwip_platform}/lwipopts-rs911x.h",
