+++ src/platform/silabs/efr32/args.gni
@@ -20,9 +20,9 @@ import("${chip_root}/examples/platform/silabs/args.gni")
 import("${chip_root}/src/crypto/crypto.gni")
 
 # ARM architecture flags will be set based on silabs_family.
-arm_platform_config = "${efr32_sdk_build_root}/silabs_arm.gni"
+arm_platform_config = "${silabs_sdk_build_root}/silabs_arm.gni"
 
-mbedtls_target = "${efr32_sdk_build_root}:efr32_sdk"
+mbedtls_target = "${silabs_sdk_build_root}:silabs_sdk"
 
 openthread_external_mbedtls = mbedtls_target
