+++ src/platform/silabs/wifi_args.gni
@@ -21,9 +21,9 @@ import("${chip_root}/examples/platform/silabs/args.gni")
 import("${chip_root}/src/crypto/crypto.gni")
 
 # ARM architecture flags will be set based on silabs_family.
-arm_platform_config = "${efr32_sdk_build_root}/silabs_arm.gni"
+arm_platform_config = "${silabs_sdk_build_root}/silabs_arm.gni"
 
-mbedtls_target = "${efr32_sdk_build_root}:efr32_sdk"
+mbedtls_target = "${silabs_sdk_build_root}:silabs_sdk"
 
 # default to platform crypto implementation but allow commandline override
 if (chip_crypto == "") {
@@ -35,7 +35,7 @@ if (chip_crypto == "") {
 chip_use_transitional_commissionable_data_provider = false
 
 # Use GSDK lwip instead of CHIP
-lwip_root = "${efr32_sdk_build_root}/silabs_lwip"
+lwip_root = "${silabs_sdk_build_root}/silabs_lwip"
 
 #lwip_platform = "external"
 lwip_platform = "silabs"
